Support PHP CodeSniffer 3.x
Using guide: https://github.com/squizlabs/PHP_CodeSniffer/wiki/Version-3.0-Upgrade-Guide

Update to work with PHP CodeSniffer 2.x and 3.x

Move non-sniffs to a different namespace

CodeSniffer 3.x will assume all classes in the
VariableAnalysis\Sniffs\VariableAnalysis namespace are sniffs.
