+++ VariableAnalysis/Scope/ScopeInfo.php
@@ -1,0 +1,36 @@
+<?php
+/**
+ * This file is part of the VariableAnalysis addon for PHP_CodeSniffer.
+ *
+ * PHP version 5
+ *
+ * @category  PHP
+ * @package   PHP_CodeSniffer
+ * @author    Sam Graham <php-codesniffer-variableanalysis BLAHBLAH illusori.co.uk>
+ * @copyright 2011-2012 Sam Graham <php-codesniffer-variableanalysis BLAHBLAH illusori.co.uk>
+ * @license   http://www.opensource.org/licenses/bsd-license.php BSD License
+ * @link      http://pear.php.net/package/PHP_CodeSniffer
+ */
+
+namespace VariableAnalysis\Scope;
+
+/**
+ * Holds details of a scope.
+ *
+ * @category  PHP
+ * @package   PHP_CodeSniffer
+ * @author    Sam Graham <php-codesniffer-variableanalysis BLAHBLAH illusori.co.uk>
+ * @copyright 2011-2012 Sam Graham <php-codesniffer-plugins BLAHBLAH illusori.co.uk>
+ * @link      http://pear.php.net/package/PHP_CodeSniffer
+ */
+class ScopeInfo {
+    public $owner;
+    public $opener;
+    public $closer;
+    public $variables = array();
+
+    function __construct($currScope) {
+        $this->owner = $currScope;
+        // TODO: extract opener/closer
+    }
+}
